Switched layout order to demonstrate field packing & alignment.

+++ native/src/main/cpp/main.h
@@ -5,8 +5,8 @@ extern "C" {
     int32_t add_short(int32_t a, int16_t b);
 
     typedef struct {
-        int32_t a;
-        int16_t b;
+        int16_t a;
+        int32_t b;
     } numbers_t;
 
     int32_t add_numbers_t(numbers_t* n);
